Refactor dependencies, create service module, convert NavItem to a Kotlin data class.
Dependencies in the Gradle build files were replaced from direct declaration to use of version catalogs declared in "dependencyResolutionManagement". This provides a single source of truth for library versions and simplifies upgrades.

A new service module was created to streamline organization and further separate concerns.

NavItem class was converted to a Kotlin data class.